replace normalization in core to support batches of examples with different sequence lengths
